requirements.txt
openai
groq
wave
keyboard
pyaudio
python-dotenv
